# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD Laravel ",
        "description": "API para administrar multimples servicios y relaciones entre tablas",
        "contact": {
            "email": "jhonfredyaya04@gmail.com"
        },
        "license": {
            "name": "Repocitorio",
            "url": "https://github.com/JhonFac/Crud_Laravel.git"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/page/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtiene todos los usuarios",
                "description": "Display a listing of the resource.",
                "operationId": "6245d067fd87883931cf376056b10006",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de usuarios"
                    }
                }
            }
        },
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crea un nuevo usuario",
                "description": "Store a newly created resource in storage.",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "image_path": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtiene el usuario por ID",
                "description": "Display the specified resource.",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Numero de la pagina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de usuarios"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Actualiza un usuario existente",
                "description": "Update the specified resource in storage.",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "image_path": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Elimina un usuario existente",
                "description": "Remove the specified resource from storage.",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
...
}